Split core into files and options scripts
